change int to double in record and fix some bugs

+++ app/src/main/java/com/example/jibi/models/SummaryMoney.kt
@@ -1,9 +1,9 @@
 package com.example.jibi.models
 
 data class SummaryMoney(
-    var balance: Int = 0,
-    var income: Int = 0,
-    var expenses: Int = 0
+    var balance: Double = 0.0,
+    var income: Double = 0.0,
+    var expenses: Double = 0.0
 )
 //extra
 /*
